Wire-mapper refactor, Android token backup exclusion, and API consistency fixes (#26)
* Refactored data classes and added mappers for response to domain model

* Added backup exclusion for the stored token in android

* Addressed few code inconsistencies

* Added few tests

* Test update after rebasing

* added netwrok logging gate

* addressed review comments from @utkrisht

+++ auth0-core/src/commonMain/kotlin/com/auth0/kmp/core/credentials/CredentialsManager.kt
@@ -30,7 +30,7 @@ public interface CredentialsManager : AutoCloseable {
      * @param minTtl the minimum remaining lifetime, in seconds, the access token
      *   must have to be considered valid.
      */
-    public suspend fun hasValidCredentials(minTtl: Long = 0): Boolean
+    public suspend fun hasValidCredentials(minTtl: Int = 0): Boolean
 
     /**
      * Returns the stored credentials, renewing them first when required.
